+++ src/Elmish/React/DOM.js
@@ -1,1 +1,18 @@
-exports.fragment_ = require("react").Fragment
+const React = require("react")
+const ReactDOM = require("react-dom")
+
+
+exports.fragment_ = React.Fragment
+
+exports.render = function(reactElement) {
+  return function(domElement) {
+    return function() {
+      ReactDOM.render(reactElement, domElement)
+    }
+  }
+}
+exports.unmountComponentAtNode = function(domElement) {
+  return function() {
+    ReactDOM.unmountComponentAtNode(domElement)
+  }
+}